pyproject.toml
[tool.poetry]
name = "analyst"
version = "1.0.0"
description = "A bot"
authors = ["ashwinvin <ashwinvinodsa@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.6.0"}
coloredlogs = "^15.0"
jishaku = "^1.20.0"
requests = "^2.25.1"
asyncpg = "^0.22.0"
wavelink = "^0.9.9"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
python-dotenv = "^0.15.0"
samp-client = "^3.0.1"
beautifultable = "^1.0.1"
pre-commit = "^2.12.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
